repetidor/pyproject.toml
[tool.poetry]
name = "repl-nix-bash"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "repl_nix_bash"}]

[tool.poetry.dependencies]
python = "^3.10"
streamlink = "^6.4.2"
redis = "^5.0.1"
python-memcached = "^1.59"
livestreamer = "^1.12.2"
yt-dlp = "2023.12.30"
vlc-http = "0.9.0"
vlc-controller = "0.0.6"
vlc-pypackage = "0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
